+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### 2.14.4
 
+- Fix it should not show duplicated PN when changing phone index (issue 873)
 - Fix it should reset PhoneAppli getProductInfo if logout (issue 868)
 - Fix it should reset navigation PhoneAppli if update account to a new one (issue 867)